finish l2vpn type helpers
