+++ cosmo/l2vpnhelpertypes.py
@@ -1,9 +1,12 @@
+import ipaddress
 import warnings
 from abc import abstractmethod, ABCMeta
 from functools import singledispatchmethod
 
+from cosmo.abstractroutervisitor import AbstractRouterExporterVisitor
 from cosmo.common import head
-from cosmo.types import InterfaceType, VLANType, AbstractNetboxType, DeviceType
+from cosmo.types import InterfaceType, VLANType, AbstractNetboxType, DeviceType, L2VPNType, CosmoLoopbackType, \
+    L2VPNTerminationType
 
 
 # FIXME simplify this!
@@ -54,7 +57,13 @@ class VlanBridgeEncapTrait(AbstractEncapTrait, metaclass=ABCMeta):
         return "vlan-bridge"
 
 
-class AbstractL2VpnType(metaclass=ABCMeta):
+# FIXME simplify this!
+class AbstractL2vpnTypeTerminationVisitor(AbstractRouterExporterVisitor, metaclass=ABCMeta):
+    def __init__(self, *args, loopbacks_by_device: dict[str, CosmoLoopbackType], asn: int, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.loopbacks_by_device = loopbacks_by_device
+        self.asn = asn
+
     @staticmethod
     @abstractmethod
     def getNetboxTypeName() -> str:
@@ -81,7 +90,7 @@ class AbstractL2VpnType(metaclass=ABCMeta):
         return f"{cls.getNetboxTypeName().upper()}: {i} is not a valid number of terminations, ignoring..."
 
 
-    def getAssociatedEncapType(self, o: type[AbstractNetboxType]) -> str|None:
+    def getAssociatedEncapType(self, o: AbstractNetboxType) -> str|None:
         associated_encap = head(list(filter(
             lambda encap: encap is not None,
             [t().accept(o) for t in self.getAssociatedEncapTraits()]
@@ -98,8 +107,43 @@ class AbstractL2VpnType(metaclass=ABCMeta):
         warnings.warn(f"{self.getNetboxTypeName()} L2VPN does not support {type(o)} terminations.")
         return
 
+    def spitInterfaceEncapFor(self, o: VLANType|InterfaceType):
+        encap_type = self.getAssociatedEncapType(o)
+        if encap_type == "ethernet-ccc" and isinstance(o, InterfaceType) and o.isSubInterface():
+            return {  # ethernet-ccc is on physical interface
+                self._interfaces_key: {
+                    o.getSubInterfaceParentInterfaceName(): {
+                        "encapsulation": encap_type
+                    }
+                }
+            }
+        elif isinstance(o, InterfaceType): # other types are on virtual interface
+            return {
+                self._interfaces_key: {
+                    **o.spitInterfacePathWith({
+                        "encapsulation": encap_type
+                    })
+                }
+            }
+        elif isinstance(o, VLANType):
+            linked_interface = head(
+                list(
+                    filter(
+                        lambda i: i.getParent(DeviceType) == o.getParent(DeviceType),
+                        o.getInterfacesAsTagged()
+                    )
+                )
+            )
+            return {
+                self._interfaces_key: {
+                    **linked_interface.spitInterfacePathWith({
+                        "encapsulation": encap_type
+                    })
+                }
+            }
 
-class AbstractP2PL2VpnType(AbstractL2VpnType, metaclass=ABCMeta):
+
+class AbstractP2PL2VpnTypeTerminationVisitor(AbstractL2vpnTypeTerminationVisitor, metaclass=ABCMeta):
     _p2p_authorized_terminations_n = 2
 
     def isValidNumberOfTerminations(self, i: int):
@@ -112,7 +156,7 @@ class AbstractP2PL2VpnType(AbstractL2VpnType, metaclass=ABCMeta):
                 f"{super().getInvalidNumberOfTerminationsErrorMessage(i)}")
 
 
-class AbstractAnyToAnyL2VpnType(AbstractL2VpnType, metaclass=ABCMeta):
+class AbstractAnyToAnyL2VpnTypeTerminationVisitor(AbstractL2vpnTypeTerminationVisitor, metaclass=ABCMeta):
     _any_to_any_min_terminations_n = 2
 
     def isValidNumberOfTerminations(self, i: int):
@@ -125,7 +169,43 @@ class AbstractAnyToAnyL2VpnType(AbstractL2VpnType, metaclass=ABCMeta):
                 f"{super().getInvalidNumberOfTerminationsErrorMessage(i)}")
 
 
-class EPLL2VpnType(AbstractP2PL2VpnType):
+#                  v so that it does not appear in subclasses of any2any or p2p since we're enumerating from there
+class EPLEVPLL2vpnTypeTraits(AbstractL2vpnTypeTerminationVisitor, metaclass=ABCMeta):
+    def processInterfaceTypeTermination(self, o: InterfaceType):
+        parent_l2vpn = o.getParent(L2VPNType)
+        local = next(filter(
+            lambda i: (
+                    isinstance(i, InterfaceType)
+                    and i == o
+            ),
+            parent_l2vpn.getTerminations()
+        ))
+        remote = next(filter(
+            lambda i: i != local,
+            parent_l2vpn.getTerminations()
+        ))
+        # + l2circuits
+        remote_end_loopback = self.loopbacks_by_device.get(remote.getAssociatedDevice().getName())
+        return {
+            self._l2circuits_key: {
+                parent_l2vpn.getName().replace("WAN: ", ""): {
+                    "interfaces": {
+                        o.getName(): {
+                            "local_label": 1_000_000 + int(local.getParent(L2VPNTerminationType).getID()),
+                            "remote_label": 1_000_000 + int(remote.getParent(L2VPNTerminationType).getID()),
+                            "remote_ip": str(ipaddress.ip_interface(remote_end_loopback.getIpv4()).ip),
+                        }
+                    },
+                    "description": f"{parent_l2vpn.getType().upper()}: "
+                                   f"{parent_l2vpn.getName().replace('WAN: ', '')} "
+                                   f"via {remote.getAssociatedDevice().getName()}",
+                }
+            }
+        } | self.spitInterfaceEncapFor(o)
+
+
+# for MRO, common traits need to be 1st
+class EPLL2VpnTypeTerminationVisitor(EPLEVPLL2vpnTypeTraits, AbstractP2PL2VpnTypeTerminationVisitor):
     @staticmethod
     def getAssociatedEncapTraits() -> list[type[AbstractEncapTrait]]:
         return [ # order is important!
@@ -142,10 +222,11 @@ class EPLL2VpnType(AbstractP2PL2VpnType):
         return InterfaceType
 
     def processInterfaceTypeTermination(self, o: InterfaceType):
-        pass
+        super().processInterfaceTypeTermination(o)
 
 
-class EVPLL2VpnType(AbstractP2PL2VpnType):
+# for MRO, common traits need to be 1st
+class EVPLL2VpnTypeTerminationVisitor(EPLEVPLL2vpnTypeTraits, AbstractP2PL2VpnTypeTerminationVisitor):
     @staticmethod
     def getAssociatedEncapTraits() -> list[type[AbstractEncapTrait]]:
         return [
@@ -164,10 +245,10 @@ class EVPLL2VpnType(AbstractP2PL2VpnType):
         return InterfaceType
 
     def processInterfaceTypeTermination(self, o: InterfaceType) -> dict | None:
-        pass
+        return super().processInterfaceTypeTermination(o)
 
 
-class VPWSL2VpnType(AbstractP2PL2VpnType):
+class VPWSL2VpnTypeTerminationVisitor(AbstractP2PL2VpnTypeTerminationVisitor):
     @staticmethod
     def getAssociatedEncapTraits() -> list[type[AbstractEncapTrait]]:
         return [
@@ -186,10 +267,46 @@ class VPWSL2VpnType(AbstractP2PL2VpnType):
         return InterfaceType
 
     def processInterfaceTypeTermination(self, o: InterfaceType) -> dict | None:
-        pass
+        parent_l2vpn = o.getParent(L2VPNType)
+        # find local end
+        local = next(filter(
+            lambda i: (
+                    isinstance(i, InterfaceType)
+                    and i.getAssociatedDevice() == o.getParent(DeviceType)
+            ),
+            parent_l2vpn.getTerminations()
+        ))
+        # remote end is the other one
+        remote = next(filter(
+            lambda i: i != local,
+            parent_l2vpn.getTerminations()
+        ))
+        return {
+            self._vrf_key: {
+                parent_l2vpn.getName().replace("WAN: ", ""): {
+                    "interfaces": [o.getName()],
+                    "description": f"VPWS: {parent_l2vpn.getName().replace('WAN: VS_', '')}",
+                    "instance_type": "evpn-vpws",
+                    "route-distinguisher": f"{self.asn}:{str(parent_l2vpn.getIdentifier())}",
+                    "vrf-target": f"target:1:{str(parent_l2vpn.getIdentifier())}",
+                    "protocols": {
+                        "evpn": {
+                            "interfaces": {
+                                o.getName(): {
+                                    "vpws_service_id": {
+                                        "local": int(local.getID()),
+                                        "remote": int(remote.getID()),
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        } | self.spitInterfaceEncapFor(o)
 
 
-class VXLANEVPNL2VpnType(AbstractAnyToAnyL2VpnType):
+class VXLANEVPNL2VpnTypeTerminationVisitor(AbstractAnyToAnyL2VpnTypeTerminationVisitor):
     @staticmethod
     def getAssociatedEncapTraits() -> list[type[AbstractEncapTrait]]:
         return [ VlanBridgeEncapTrait ]
@@ -205,10 +322,44 @@ class VXLANEVPNL2VpnType(AbstractAnyToAnyL2VpnType):
         return VLANType
 
     def processVLANTypeTermination(self, o: VLANType) -> dict | None:
-        pass
+        def spitNameForInterfaces(int_or_vlan: InterfaceType | VLANType):
+            if isinstance(int_or_vlan, InterfaceType):
+                return int_or_vlan.getName()
+        parent_l2vpn = o.getParent(L2VPNType)
+        vlan_id = None
+        if o.getUntaggedVLAN():
+            vlan_id = o.getUntaggedVLAN().getVID()
+        return {
+            self._vrf_key: {
+                parent_l2vpn.getName().replace("WAN: ", ""): {
+                    "description": f"Virtual Switch {parent_l2vpn.getName().replace('WAN: VS_', '')}",
+                    "instance-type": "virtual-switch",
+                    "route_distinguisher": f"{self.asn}:{str(parent_l2vpn.getIdentifier())}",
+                    "vrf_target": f"target:1:{str(parent_l2vpn.getIdentifier())}",
+                    "protocols": {
+                        "evpn": {
+                            "vni": [parent_l2vpn.getIdentifier()]
+                        }
+                    },
+                    "bridge_domains": [
+                        {
+                            # I have to do this, otherwise deepmerge will assume non-uniqueness and we'll have
+                            # duplicated bridge_domains. this is because we're within a dict within a list.
+                            "interfaces": [spitNameForInterfaces(iov) for iov in parent_l2vpn.getTerminations()],
+                            "vlan_id": vlan_id,
+                            "name": parent_l2vpn.getName(),
+                            "vxlan": {
+                                "ingress_node_replication": True,
+                                "vni": parent_l2vpn.getIdentifier(),
+                            }
+                        }
+                    ]
+                }
+            }
+        } | self.spitInterfaceEncapFor(o)
 
 
-class MPLSEVPNL2VpnType(AbstractAnyToAnyL2VpnType):
+class MPLSEVPNL2VpnTypeTerminationVisitor(AbstractAnyToAnyL2VpnTypeTerminationVisitor):
     @staticmethod
     def getAssociatedEncapTraits() -> list[type[AbstractEncapTrait]]:
         return [ VlanBridgeEncapTrait ]
@@ -224,7 +375,22 @@ class MPLSEVPNL2VpnType(AbstractAnyToAnyL2VpnType):
         return InterfaceType, VLANType
 
     def processInterfaceTypeTermination(self, o: InterfaceType) -> dict | None:
-        pass
+        parent_l2vpn = o.getParent(L2VPNType)
+        return {
+            self._vrf_key: {
+                parent_l2vpn.getName().replace("WAN: ", ""): {
+                    "interfaces": [o.getName()],
+                    "description": f"MPLS-EVPN: {parent_l2vpn.getName().replace('WAN: VS_', '')}",
+                    "instance_type": "evpn",
+                    "protocols": {
+                        "evpn": {},
+                    },
+                    "route_distinguisher": f"{self.asn}:{str(parent_l2vpn.getIdentifier())}",
+                    "vrf_target": f"target:1:{str(parent_l2vpn.getIdentifier())}",
+                }
+            }
+        } | self.spitInterfaceEncapFor(o)
 
     def processVLANTypeTermination(self, o: VLANType) -> dict | None:
+        # TODO!!!
         pass
